3: Изменить MealRepository и InMemoryMealRepository таким образом, чтобы вся еда всех пользователей находилась в одном общем хранилище, но при этом каждый конкретный авторизованный пользователь мог видеть и редактировать только свою еду.
3.1: реализовать хранение еды для каждого пользователя можно с добавлением поля userId в Meal ИЛИ без него (как нравится). Напомню, что репозиторий один и приложение может работать одновременно с многими пользователями.

3.2: если по запрошенному id еда отсутствует или чужая, возвращать null/false (см. комментарии в UserRepository)

3.3: список еды возвращать отсортированный в обратном порядке по датам

3.4: атомарность операций не требуется (коллизии при одновременном изменении еды одного пользователя можно не учитывать)

Типичные ошибки и подсказки по реализации:
2. В реализации InMemoryMealRepository
2.1: В Meal, которая приходит из браузера в контроллер, нет никакой информации о пользователе (еда приходит в контроллер БЕЗ user/userId, она может быть только от авторизированного пользователя). По id еды и авторизованному пользователю нужно проверить ее принадлежность (его это еда или чужая) Проверьте сценарий: авторизованный пользователь пробует изменить чужую еду (id еды ему не принадлежит).
2.2: get\update\delete - следите, чтобы не было NPE (NullPointException может быть, если в хранилище отсутствует юзер или еда).
2.3: Фильтрацию по датам сделать в репозитории т.к. из базы будем брать сразу отфильтрованные по дням записи. Следите чтобы первый и последний день не были обрезаны, иначе сумма калорий будет неверная.
2.4: Если запрашивается список и он пустой, не возвращайте NULL! По пустому списку можно легко итерироваться без риска NullPoinException.
2.5: Не дублируйте код в getAll и метод с фильтрацией
2.6: попробуйте учесть, что следующая реализация (сортировка, фильтрация) будет делаться прямо в базе данных

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -3,11 +3,12 @@ package ru.javawebinar.topjava.repository.inmemory;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
+import ru.javawebinar.topjava.web.SecurityUtil;
 
-import java.util.Collection;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
 public class InMemoryMealRepository implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
@@ -30,17 +31,27 @@ public class InMemoryMealRepository implements MealRepository {
 
     @Override
     public boolean delete(int id) {
-        return repository.remove(id) != null;
+        Meal meal = repository.get(id);
+        if (meal != null && meal.getUserId() == SecurityUtil.authUserId()) {
+            return repository.remove(id) != null;
+        }
+        return false;
     }
 
     @Override
     public Meal get(int id) {
-        return repository.get(id);
+        Meal meal = repository.get(id);
+        if (meal != null && meal.getUserId() == SecurityUtil.authUserId()) {
+            return meal;
+        }
+        return null;
     }
 
     @Override
     public Collection<Meal> getAll() {
-        return repository.values();
+        List<Meal> meals = repository.values().stream().filter(meal -> meal.getUserId() == SecurityUtil.authUserId())
+                .sorted((x,y) -> y.getDateTime().compareTo(x.getDateTime())).collect(Collectors.toList());
+        return !meals.isEmpty() ? meals : Collections.emptyList();
     }
 }
